Fix bugs PromoteI1ToI8Pass and move it to common input conversion.
Deletes inserting the op right before the first use of the constant.
Generating the op right after the constant guarantees the dominitio
because the constant already meets it. Plus, a constant could be used in
different blocks. In this context, it will crash in the check.

Also adds the pass to MHLO pipeline.
