+++ iree/compiler/Dialect/Flow/Transforms/CMakeLists.txt
@@ -49,7 +49,6 @@ iree_cc_library(
     "PadTensorToSubTensorInsert.cpp"
     "PassDetail.h"
     "Passes.cpp"
-    "PromoteI1ToI8Pass.cpp"
     "StripAndSplatConstantVariables.cpp"
     "TypeConverter.cpp"
     "VerifyInputLegality.cpp"
