+++ iree/compiler/InputConversion/Common/test/promote_i1_to_i8.mlir
@@ -1,5 +1,4 @@
-
-// RUN: iree-opt -split-input-file -pass-pipeline='builtin.func(iree-flow-promote-i1-to-i8)' %s | IreeFileCheck %s
+// RUN: iree-opt -split-input-file -pass-pipeline='builtin.func(iree-promote-i1-to-i8)' %s | IreeFileCheck %s
 
 // CHECK: #[[$MAP:.+]] = affine_map<(d0) -> (d0)>
